feat(fastify): authentication
