+++ turbo.json
@@ -72,7 +72,8 @@
         "BETTER_AUTH_TRUSTED_ORIGINS",
         "RESEND_API_KEY",
         "EMAIL_FROM",
-        "EMAIL_FROM_NAME"
+        "EMAIL_FROM_NAME",
+        "VITEST_WORKER_ID"
       ]
     },
     "lint:eslint:fix": {
@@ -97,7 +98,8 @@
         "BETTER_AUTH_TRUSTED_ORIGINS",
         "RESEND_API_KEY",
         "EMAIL_FROM",
-        "EMAIL_FROM_NAME"
+        "EMAIL_FROM_NAME",
+        "VITEST_WORKER_ID"
       ]
     },
     "checktypes": {
@@ -126,7 +128,8 @@
         "EMAIL_FROM_NAME",
         "BETTER_AUTH_SECRET",
         "BETTER_AUTH_URL",
-        "BETTER_AUTH_TRUSTED_ORIGINS"
+        "BETTER_AUTH_TRUSTED_ORIGINS",
+        "VITEST_WORKER_ID"
       ]
     },
     "dev": {
